feat: source cleanup
