+++ src/component/atoms/Text.tsx
@@ -1,17 +1,16 @@
 import React from "react";
 import { Platform, Text as RNText, TextProps } from "react-native";
-import { AppColors } from "~/styles/colors";
-import { font, FONT_LINE_HEIGHT, FONT_SIZE, FONT_WEIGHT, setFontStyle } from "~/styles/font";
+import { font, FontStyle, setFontStyle } from "~/styles/font";
 import { useThemeStyle } from "~/styles/theme";
 
-export const Text: React.FC<
-	TextProps & {
-		size?: FONT_SIZE;
-		lineHeight?: FONT_LINE_HEIGHT;
-		weight?: FONT_WEIGHT;
-		color?: AppColors;
-	}
-> = ({ style, size = "L", weight = "Regular", lineHeight = "Deafult", color, ...props }) => {
+export const Text: React.FC<TextProps & FontStyle> = ({
+	style,
+	size = "L",
+	weight = "Regular",
+	lineHeight = "Deafult",
+	color,
+	...props
+}) => {
 	const theme = useThemeStyle();
 	return (
 		<RNText
